@@ -2,169 +2,125 @@
   "name": "silhouette-play-react-seed",
   "version": "4.0.0",
   "description": "Silhouette Play React Seed Template",
-  "license": "MIT",
+  "license": "Apache",
   "main": "index.js",
   "engines": {
-    "node": ">=4.5.0",
+    "node": ">=6.9.0",
     "npm": "^3.0.0"
   },
-  "scripts": {
-    "clean": "rimraf dist",
-    "compile": "better-npm-run compile",
-    "lint": "eslint --ext .js --ext .jsx bin build config server src tests",
-    "lint:fix": "npm run lint -- --fix",
-    "start": "better-npm-run start",
-    "dev": "better-npm-run dev",
-    "test": "better-npm-run test",
-    "test:dev": "npm run test -- --watch",
-    "deploy": "better-npm-run deploy",
-    "deploy:dev": "better-npm-run deploy:dev",
-    "deploy:prod": "better-npm-run deploy:prod",
-    "codecov": "cat coverage/*/lcov.info | codecov",
-    "flow": "flow check || exit 0",
-    "flow-typed:install": "flow-typed install || exit 0"
-  },
-  "betterScripts": {
-    "compile": {
-      "command": "node bin/compile",
-      "env": {
-        "DEBUG": "app:*"
+  "neutrino": {
+    "use": [
+      "neutrino-options.js",
+      "neutrino-preset-react",
+      "neutrino-preset-karma",
+      "neutrino-webpack.js",
+      "neutrino-sass.js",
+      "neutrino-karma.js"
+    ],
+    "config": {
+      "entry": {
+        "vendor": [
+          "flux-standard-action",
+          "js-cookie",
+          "lodash",
+          "prop-types",
+          "react",
+          "react-bootstrap",
+          "react-dom",
+          "react-redux",
+          "react-redux-form",
+          "react-router",
+          "react-s-alert",
+          "redux",
+          "redux-actions",
+          "redux-auth-wrapper",
+          "redux-saga",
+          "redux-thunk",
+          "validator",
+          "whatwg-fetch"
+        ]
       }
     },
-    "dev": {
-      "command": "nodemon bin/dev-server --ignore dist --ignore coverage --ignore tests --ignore src",
-      "env": {
-        "NODE_ENV": "development",
-        "DEBUG": "app:*"
-      }
-    },
-    "deploy": {
-      "command": "npm run clean && npm run compile",
-      "env": {
-        "DEBUG": "app:*"
-      }
-    },
-    "deploy:dev": {
-      "command": "npm run deploy",
-      "env": {
-        "NODE_ENV": "development",
-        "DEBUG": "app:*"
-      }
-    },
-    "deploy:prod": {
-      "command": "npm run deploy",
-      "env": {
-        "NODE_ENV": "production",
-        "DEBUG": "app:*"
-      }
-    },
-    "start": {
-      "command": "node bin/dev-server",
-      "env": {
-        "DEBUG": "app:*"
-      }
-    },
-    "test": {
-      "command": "node ./node_modules/karma/bin/karma start config/karma.config",
-      "env": {
-        "NODE_ENV": "test",
-        "DEBUG": "app:*"
+    "options": {
+      "html": {
+        "title": "Silhouette Play React Seed Template"
       }
     }
   },
+  "scripts": {
+    "clean": "rimraf ../target/npm",
+    "start": "neutrino start",
+    "build": "neutrino build",
+    "test": "neutrino test",
+    "lint": "eslint --ext .js,.jsx src/ test/",
+    "lint:fix": "npm run lint -- --fix",
+    "flow": "flow check || exit 0",
+    "flow-typed:install": "flow-typed install || exit 0"
+  },
   "dependencies": {
-    "babel-core": "^6.17.0",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-0": "^6.3.13",
-    "babel-runtime": "^6.11.6",
-    "better-npm-run": "0.0.13",
-    "bootstrap-sass": "^3.3.7",
-    "compression": "^1.6.2",
-    "css-loader": "^0.26.0",
-    "cssnano": "^3.7.4",
-    "debug": "^2.2.0",
-    "eslint-plugin-import": "^2.2.0",
-    "extract-text-webpack-plugin": "^1.0.0",
-    "fetch-mock": "^5.5.0",
-    "file-loader": "^0.9.0",
-    "flux-standard-action": "^1.0.0",
-    "fs-extra": "^1.0.0",
-    "html-webpack-plugin": "^2.22.0",
-    "imports-loader": "^0.7.0",
-    "ip": "^1.1.2",
+    "flux-standard-action": "^1.2.0",
     "js-cookie": "^2.1.4",
-    "json-loader": "^0.5.4",
-    "lodash": "^4.16.6",
-    "node-sass": "^4.0.0",
-    "normalize.css": "^5.0.0",
-    "postcss-loader": "^1.1.0",
+    "lodash": "^4.17.4",
+    "normalize.css": "^7.0.0",
+    "prop-types": "^15.5.10",
     "react": "^15.0.0",
-    "react-bootstrap": "^0.30.7",
+    "react-bootstrap": "^0.31.0",
     "react-dom": "^15.0.0",
-    "react-icons": "^2.2.1",
-    "react-redux": "^5.0.1",
-    "react-redux-form": "^1.8.1",
+    "react-icons": "^2.2.5",
+    "react-redux": "^5.0.4",
+    "react-redux-form": "^1.11.0",
     "react-router": "^3.0.0",
-    "react-s-alert": "^1.2.2",
+    "react-s-alert": "^1.3.0",
     "redux": "^3.6.0",
-    "redux-actions": "^0.13.0",
-    "redux-auth-wrapper": "^1.0.0",
-    "redux-saga": "^0.15.1",
-    "redux-thunk": "^2.0.0",
-    "rimraf": "^2.5.4",
-    "sass-loader": "^4.0.0",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.6",
+    "redux-actions": "^2.0.3",
+    "redux-auth-wrapper": "^1.1.0",
+    "redux-saga": "^0.15.3",
+    "redux-thunk": "^2.2.0",
     "validator": "^7.0.0",
-    "webpack": "^1.12.14",
-    "whatwg-fetch": "^2.0.1",
-    "yargs": "^6.3.0"
+    "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
-    "babel-eslint": "^7.1.0",
-    "babel-plugin-istanbul": "^3.0.0",
-    "babel-polyfill": "^6.20.0",
+    "babel-eslint": "^7.2.3",
+    "babel-plugin-istanbul": "^4.1.3",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "bootstrap-sass": "^3.3.7",
     "chai": "^3.4.1",
     "chai-as-promised": "^6.0.0",
     "chai-enzyme": "^0.6.1",
     "chai-shallow-deep-equal": "^1.4.6",
     "cheerio": "^0.22.0",
-    "codecov": "^1.0.1",
-    "connect-history-api-fallback": "^1.3.0",
+    "css-loader": "^0.28.0",
     "dirty-chai": "^1.2.2",
     "enzyme": "^2.0.0",
-    "eslint": "^3.0.1",
-    "eslint-config-airbnb": "^12.0.0",
-    "eslint-config-standard": "^6.0.0",
-    "eslint-config-standard-react": "^4.0.0",
-    "eslint-import-resolver-webpack": "^0.7.1",
-    "eslint-plugin-babel": "^4.0.0",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-import-resolver-webpack": "^0.8.1",
     "eslint-plugin-flowtype": "^2.30.3",
-    "eslint-plugin-jsx-a11y": "^2.2.2",
-    "eslint-plugin-promise": "^3.0.0",
-    "eslint-plugin-react": "^6.0.0",
-    "eslint-plugin-standard": "^2.0.0",
-    "express": "^4.14.0",
-    "flow-bin": "^0.41.0",
-    "flow-typed": "^2.0.0",
-    "karma": "^1.0.0",
-    "karma-coverage": "^1.0.0",
-    "karma-mocha": "^1.0.1",
-    "karma-mocha-reporter": "^2.0.0",
-    "karma-phantomjs-launcher": "^1.0.2",
-    "karma-webpack-with-fast-source-maps": "^1.9.2",
-    "mocha": "^3.0.1",
-    "nodemon": "^1.10.2",
-    "phantomjs-prebuilt": "^2.1.12",
-    "react-addons-test-utils": "^15.0.0",
-    "redbox-react": "^1.2.10",
-    "redux-saga-test-plan": "^3.0.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "4.0.0",
+    "eslint-plugin-react": "^6.9.0",
+    "fetch-mock": "^5.5.0",
+    "flow-bin": "^0.46.0",
+    "flow-typed": "^2.1.2",
+    "karma": "^1.7.0",
+    "karma-phantomjs-launcher": "^1.0.4",
+    "neutrino": "^5.8.0",
+    "neutrino-preset-karma": "^5.8.0",
+    "neutrino-preset-react": "^5.8.0",
+    "node-sass": "^4.5.2",
+    "phantomjs-prebuilt": "^2.1.14",
+    "react-addons-test-utils": "^15.5.1",
+    "react-hot-loader": "next",
+    "react-test-renderer": "^15.5.4",
+    "redbox-react": "^1.3.6",
+    "redux-saga-test-plan": "^3.0.2",
+    "rimraf": "^2.6.1",
+    "sass-loader": "^6.0.3",
     "sinon": "^1.17.5",
     "sinon-chai": "^2.8.0",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-hot-middleware": "^2.12.2"
+    "style-loader": "^0.16.1",
+    "webpack": "^2.5.1"
   }
 }
