@@ -1,5 +1,5 @@
-// flow-typed signature: d1937428a474c8cdc6e8c39a336d2972
-// flow-typed version: <<STUB>>/babel-eslint_v^7.1.0/flow_v0.41.0
+// flow-typed signature: f3c3dd3c4bea413962ad456458ec0804
+// flow-typed version: <<STUB>>/babel-eslint_v^7.2.3/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
