@@ -1,5 +1,5 @@
-// flow-typed signature: fd8f5c791d722250223b647f24608936
-// flow-typed version: <<STUB>>/babel-plugin-istanbul_v^3.0.0/flow_v0.41.0
+// flow-typed signature: c1571149f63879feef620cd300d2d36a
+// flow-typed version: <<STUB>>/babel-plugin-istanbul_v^4.1.3/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
