@@ -1,5 +1,5 @@
-// flow-typed signature: a4f1fa054456afbac10ad1b7608f0009
-// flow-typed version: <<STUB>>/babel-polyfill_v^6.20.0/flow_v0.41.0
+// flow-typed signature: 91ca23813d845573b75a710f0e3e3619
+// flow-typed version: <<STUB>>/babel-polyfill_v^6.23.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
