@@ -1,5 +1,5 @@
-// flow-typed signature: 600d342b8a9798c7034104c07b101b49
-// flow-typed version: <<STUB>>/babel-preset-es2015_v^6.14.0/flow_v0.41.0
+// flow-typed signature: 5e71bb877ae4c3a70c181a71e2267aa6
+// flow-typed version: <<STUB>>/babel-preset-es2015_v^6.24.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
