@@ -1,5 +1,5 @@
-// flow-typed signature: 9430ebbc963afb1d5888881eca735b2a
-// flow-typed version: <<STUB>>/chai-as-promised_v^6.0.0/flow_v0.41.0
+// flow-typed signature: 6e96d666972fa2a338661801f91339e9
+// flow-typed version: <<STUB>>/chai-as-promised_v^6.0.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
