@@ -1,5 +1,5 @@
-// flow-typed signature: df84d7b66ad7da0bd2e947a71e9849b6
-// flow-typed version: <<STUB>>/chai-enzyme_v^0.6.1/flow_v0.41.0
+// flow-typed signature: 23c3006d69192dad28bd612f85de047c
+// flow-typed version: <<STUB>>/chai-enzyme_v^0.6.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
