@@ -1,5 +1,5 @@
-// flow-typed signature: 3b24a0c0d5ce795bc6c2018675a94191
-// flow-typed version: <<STUB>>/chai-shallow-deep-equal_v^1.4.6/flow_v0.41.0
+// flow-typed signature: 657b1e998017674fac1f20b1877292c4
+// flow-typed version: <<STUB>>/chai-shallow-deep-equal_v^1.4.6/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
