@@ -1,5 +1,5 @@
-// flow-typed signature: 775225030491d6dadba1c09b1be410cc
-// flow-typed version: <<STUB>>/chai_v^3.4.1/flow_v0.41.0
+// flow-typed signature: bafe8dde0fe02d1e7067b10d4e4ae513
+// flow-typed version: <<STUB>>/chai_v^3.4.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
