@@ -1,5 +1,5 @@
-// flow-typed signature: 8a9238d9b86eff86396d834e02dcf39c
-// flow-typed version: <<STUB>>/css-loader_v^0.26.0/flow_v0.41.0
+// flow-typed signature: a629747e62a1af86e3262b5b2e34c08d
+// flow-typed version: <<STUB>>/css-loader_v^0.28.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -23,6 +23,10 @@ declare module 'css-loader' {
  * needed.
  */
 declare module 'css-loader/lib/compile-exports' {
+  declare module.exports: any;
+}
+
+declare module 'css-loader/lib/createResolver' {
   declare module.exports: any;
 }
 
@@ -63,6 +67,9 @@ declare module 'css-loader/index.js' {
 }
 declare module 'css-loader/lib/compile-exports.js' {
   declare module.exports: $Exports<'css-loader/lib/compile-exports'>;
+}
+declare module 'css-loader/lib/createResolver.js' {
+  declare module.exports: $Exports<'css-loader/lib/createResolver'>;
 }
 declare module 'css-loader/lib/css-base.js' {
   declare module.exports: $Exports<'css-loader/lib/css-base'>;
