@@ -1,5 +1,5 @@
-// flow-typed signature: d88c2a0e9201cf1ceed2b488ddff8f24
-// flow-typed version: <<STUB>>/dirty-chai_v^1.2.2/flow_v0.41.0
+// flow-typed signature: d04cb5f678b5576ac9f88450efa03a7a
+// flow-typed version: <<STUB>>/dirty-chai_v^1.2.2/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
