@@ -1,5 +1,5 @@
-// flow-typed signature: 68eb60330f92ba772c0fb56e79c7ddf4
-// flow-typed version: <<STUB>>/eslint-config-airbnb_v^12.0.0/flow_v0.41.0
+// flow-typed signature: d87ff06c92fc003b43f2af6a0d084ede
+// flow-typed version: <<STUB>>/eslint-config-airbnb_v^14.1.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
