@@ -1,5 +1,5 @@
-// flow-typed signature: df09c9d2afc0a37b5e44f4925d71f54a
-// flow-typed version: <<STUB>>/eslint-import-resolver-webpack_v^0.7.1/flow_v0.41.0
+// flow-typed signature: 26be4fcfd5e2170b0e58ef04e2e45e48
+// flow-typed version: <<STUB>>/eslint-import-resolver-webpack_v^0.8.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
