@@ -1,5 +1,5 @@
-// flow-typed signature: fe9cb7ccc3d8f80df4a6f51282e0198d
-// flow-typed version: <<STUB>>/eslint-plugin-flowtype_v^2.30.3/flow_v0.41.0
+// flow-typed signature: 8165119a205d8aad7512776bfabcba3d
+// flow-typed version: <<STUB>>/eslint-plugin-flowtype_v^2.30.3/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
