@@ -1,5 +1,5 @@
-// flow-typed signature: 421c449fc02c76c59634a8f11df76eea
-// flow-typed version: <<STUB>>/eslint-plugin-import_v^2.2.0/flow_v0.41.0
+// flow-typed signature: fdc069572f193f76e11f3ff6be5fde49
+// flow-typed version: <<STUB>>/eslint-plugin-import_v^2.2.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
