@@ -1,5 +1,5 @@
-// flow-typed signature: 6cd98d5498a9db0e573d58b088fc35d6
-// flow-typed version: <<STUB>>/eslint-plugin-jsx-a11y_v^2.2.2/flow_v0.41.0
+// flow-typed signature: 6dd73c505197b0b0b4f433dd7b315a8e
+// flow-typed version: <<STUB>>/eslint-plugin-jsx-a11y_v4.0.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,11 +22,191 @@ declare module 'eslint-plugin-jsx-a11y' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
+declare module 'eslint-plugin-jsx-a11y/__mocks__/genInteractives' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__mocks__/IdentifierMock' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXAttributeMock' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXElementMock' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXExpressionContainerMock' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/__util__/parserOptionsMapper' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/index-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/accessible-emoji-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/anchor-has-content-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-activedescendant-has-tabindex-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-props-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-proptypes-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-role-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-unsupported-elements-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/click-events-have-key-events-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/heading-has-content-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/href-no-hash-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/html-has-lang-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/iframe-has-title-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/img-has-alt-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/img-redundant-alt-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-for-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/lang-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/mouse-events-have-key-events-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-access-key-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-autofocus-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-distracting-elements-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-onchange-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-redundant-roles-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-static-element-interactions-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/onclick-has-focus-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/onclick-has-role-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/role-has-required-aria-props-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/role-supports-aria-props-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/scope-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/tabindex-no-positive-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getSuggestion-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getTabIndex-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveElement-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveRole-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/parserOptionsMapper-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/schemas-test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/flow/eslint' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/lib/index' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/lib/rules/accessible-emoji' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/lib/rules/anchor-has-content' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/rules/aria-activedescendant-has-tabindex' {
   declare module.exports: any;
 }
 
@@ -62,6 +242,10 @@ declare module 'eslint-plugin-jsx-a11y/lib/rules/html-has-lang' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/lib/rules/iframe-has-title' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/lib/rules/img-has-alt' {
   declare module.exports: any;
 }
@@ -86,11 +270,19 @@ declare module 'eslint-plugin-jsx-a11y/lib/rules/no-access-key' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jsx-a11y/lib/rules/no-marquee' {
+declare module 'eslint-plugin-jsx-a11y/lib/rules/no-autofocus' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/rules/no-distracting-elements' {
   declare module.exports: any;
 }
 
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-onchange' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/rules/no-redundant-roles' {
   declare module.exports: any;
 }
 
@@ -294,11 +486,47 @@ declare module 'eslint-plugin-jsx-a11y/lib/util/isInteractiveElement' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/lib/util/isInteractiveRole' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/lib/util/schemas' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/scripts/addRuleToIndex' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/scripts/boilerplate/doc' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/scripts/boilerplate/rule' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/scripts/boilerplate/test' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/scripts/create-rule' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/src/index' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/src/rules/accessible-emoji' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/src/rules/anchor-has-content' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/src/rules/aria-activedescendant-has-tabindex' {
   declare module.exports: any;
 }
 
@@ -334,6 +562,10 @@ declare module 'eslint-plugin-jsx-a11y/src/rules/html-has-lang' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/src/rules/iframe-has-title' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/src/rules/img-has-alt' {
   declare module.exports: any;
 }
@@ -358,11 +590,19 @@ declare module 'eslint-plugin-jsx-a11y/src/rules/no-access-key' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jsx-a11y/src/rules/no-marquee' {
+declare module 'eslint-plugin-jsx-a11y/src/rules/no-autofocus' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/src/rules/no-distracting-elements' {
   declare module.exports: any;
 }
 
 declare module 'eslint-plugin-jsx-a11y/src/rules/no-onchange' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-jsx-a11y/src/rules/no-redundant-roles' {
   declare module.exports: any;
 }
 
@@ -566,116 +806,155 @@ declare module 'eslint-plugin-jsx-a11y/src/util/isInteractiveElement' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jsx-a11y/tests/index' {
+declare module 'eslint-plugin-jsx-a11y/src/util/isInteractiveRole' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/anchor-has-content' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/aria-props' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/aria-proptypes' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/aria-role' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/aria-unsupported-elements' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/click-events-have-key-events' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/heading-has-content' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/href-no-hash' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/html-has-lang' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/img-has-alt' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/img-redundant-alt' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/label-has-for' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/lang' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/mouse-events-have-key-events' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/no-access-key' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/no-marquee' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/no-onchange' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/no-static-element-interactions' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/onclick-has-focus' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/onclick-has-role' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/role-has-required-aria-props' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/role-supports-aria-props' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/scope' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/tabindex-no-positive' {
-  declare module.exports: any;
-}
-
-declare module 'eslint-plugin-jsx-a11y/tests/src/util/getSuggestion' {
+declare module 'eslint-plugin-jsx-a11y/src/util/schemas' {
   declare module.exports: any;
 }
 
 // Filename aliases
+declare module 'eslint-plugin-jsx-a11y/__mocks__/genInteractives.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/genInteractives'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__mocks__/IdentifierMock.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/IdentifierMock'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXAttributeMock.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/JSXAttributeMock'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXElementMock.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/JSXElementMock'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXExpressionContainerMock.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/JSXExpressionContainerMock'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/__util__/parserOptionsMapper.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/__util__/parserOptionsMapper'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/index-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/index-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/accessible-emoji-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/accessible-emoji-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/anchor-has-content-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/anchor-has-content-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-activedescendant-has-tabindex-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-activedescendant-has-tabindex-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-props-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-props-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-proptypes-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-proptypes-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-role-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-role-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-unsupported-elements-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-unsupported-elements-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/click-events-have-key-events-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/click-events-have-key-events-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/heading-has-content-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/heading-has-content-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/href-no-hash-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/href-no-hash-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/html-has-lang-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/html-has-lang-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/iframe-has-title-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/iframe-has-title-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/img-has-alt-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/img-has-alt-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/img-redundant-alt-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/img-redundant-alt-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-for-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-for-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/lang-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/lang-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/mouse-events-have-key-events-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/mouse-events-have-key-events-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-access-key-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-access-key-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-autofocus-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-autofocus-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-distracting-elements-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-distracting-elements-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-onchange-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-onchange-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-redundant-roles-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-redundant-roles-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-static-element-interactions-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-static-element-interactions-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/onclick-has-focus-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/onclick-has-focus-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/onclick-has-role-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/onclick-has-role-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/role-has-required-aria-props-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/role-has-required-aria-props-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/role-supports-aria-props-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/role-supports-aria-props-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/scope-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/scope-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/tabindex-no-positive-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/tabindex-no-positive-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getSuggestion-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/getSuggestion-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getTabIndex-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/getTabIndex-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveElement-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveElement-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveRole-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveRole-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/parserOptionsMapper-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/parserOptionsMapper-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/schemas-test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/schemas-test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/flow/eslint.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/flow/eslint'>;
+}
 declare module 'eslint-plugin-jsx-a11y/lib/index.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/index'>;
 }
+declare module 'eslint-plugin-jsx-a11y/lib/rules/accessible-emoji.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/accessible-emoji'>;
+}
 declare module 'eslint-plugin-jsx-a11y/lib/rules/anchor-has-content.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/anchor-has-content'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/rules/aria-activedescendant-has-tabindex.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/aria-activedescendant-has-tabindex'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/aria-props.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/aria-props'>;
@@ -701,6 +980,9 @@ declare module 'eslint-plugin-jsx-a11y/lib/rules/href-no-hash.js' {
 declare module 'eslint-plugin-jsx-a11y/lib/rules/html-has-lang.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/html-has-lang'>;
 }
+declare module 'eslint-plugin-jsx-a11y/lib/rules/iframe-has-title.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/iframe-has-title'>;
+}
 declare module 'eslint-plugin-jsx-a11y/lib/rules/img-has-alt.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/img-has-alt'>;
 }
@@ -719,11 +1001,17 @@ declare module 'eslint-plugin-jsx-a11y/lib/rules/mouse-events-have-key-events.js
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-access-key.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-access-key'>;
 }
-declare module 'eslint-plugin-jsx-a11y/lib/rules/no-marquee.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-marquee'>;
+declare module 'eslint-plugin-jsx-a11y/lib/rules/no-autofocus.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-autofocus'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/rules/no-distracting-elements.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-distracting-elements'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-onchange.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-onchange'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/rules/no-redundant-roles.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-redundant-roles'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-static-element-interactions.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-static-element-interactions'>;
@@ -875,11 +1163,38 @@ declare module 'eslint-plugin-jsx-a11y/lib/util/isHiddenFromScreenReader.js' {
 declare module 'eslint-plugin-jsx-a11y/lib/util/isInteractiveElement.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isInteractiveElement'>;
 }
+declare module 'eslint-plugin-jsx-a11y/lib/util/isInteractiveRole.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isInteractiveRole'>;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/util/schemas.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/schemas'>;
+}
+declare module 'eslint-plugin-jsx-a11y/scripts/addRuleToIndex.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/scripts/addRuleToIndex'>;
+}
+declare module 'eslint-plugin-jsx-a11y/scripts/boilerplate/doc.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/scripts/boilerplate/doc'>;
+}
+declare module 'eslint-plugin-jsx-a11y/scripts/boilerplate/rule.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/scripts/boilerplate/rule'>;
+}
+declare module 'eslint-plugin-jsx-a11y/scripts/boilerplate/test.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/scripts/boilerplate/test'>;
+}
+declare module 'eslint-plugin-jsx-a11y/scripts/create-rule.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/scripts/create-rule'>;
+}
 declare module 'eslint-plugin-jsx-a11y/src/index.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/index'>;
 }
+declare module 'eslint-plugin-jsx-a11y/src/rules/accessible-emoji.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/accessible-emoji'>;
+}
 declare module 'eslint-plugin-jsx-a11y/src/rules/anchor-has-content.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/anchor-has-content'>;
+}
+declare module 'eslint-plugin-jsx-a11y/src/rules/aria-activedescendant-has-tabindex.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/aria-activedescendant-has-tabindex'>;
 }
 declare module 'eslint-plugin-jsx-a11y/src/rules/aria-props.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/aria-props'>;
@@ -905,6 +1220,9 @@ declare module 'eslint-plugin-jsx-a11y/src/rules/href-no-hash.js' {
 declare module 'eslint-plugin-jsx-a11y/src/rules/html-has-lang.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/html-has-lang'>;
 }
+declare module 'eslint-plugin-jsx-a11y/src/rules/iframe-has-title.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/iframe-has-title'>;
+}
 declare module 'eslint-plugin-jsx-a11y/src/rules/img-has-alt.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/img-has-alt'>;
 }
@@ -923,11 +1241,17 @@ declare module 'eslint-plugin-jsx-a11y/src/rules/mouse-events-have-key-events.js
 declare module 'eslint-plugin-jsx-a11y/src/rules/no-access-key.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/no-access-key'>;
 }
-declare module 'eslint-plugin-jsx-a11y/src/rules/no-marquee.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/no-marquee'>;
+declare module 'eslint-plugin-jsx-a11y/src/rules/no-autofocus.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/no-autofocus'>;
+}
+declare module 'eslint-plugin-jsx-a11y/src/rules/no-distracting-elements.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/no-distracting-elements'>;
 }
 declare module 'eslint-plugin-jsx-a11y/src/rules/no-onchange.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/no-onchange'>;
+}
+declare module 'eslint-plugin-jsx-a11y/src/rules/no-redundant-roles.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/no-redundant-roles'>;
 }
 declare module 'eslint-plugin-jsx-a11y/src/rules/no-static-element-interactions.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/rules/no-static-element-interactions'>;
@@ -1079,81 +1403,9 @@ declare module 'eslint-plugin-jsx-a11y/src/util/isHiddenFromScreenReader.js' {
 declare module 'eslint-plugin-jsx-a11y/src/util/isInteractiveElement.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/util/isInteractiveElement'>;
 }
-declare module 'eslint-plugin-jsx-a11y/tests/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/index'>;
+declare module 'eslint-plugin-jsx-a11y/src/util/isInteractiveRole.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/util/isInteractiveRole'>;
 }
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/anchor-has-content.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/anchor-has-content'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/aria-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/aria-props'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/aria-proptypes.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/aria-proptypes'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/aria-role.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/aria-role'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/aria-unsupported-elements.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/aria-unsupported-elements'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/click-events-have-key-events.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/click-events-have-key-events'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/heading-has-content.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/heading-has-content'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/href-no-hash.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/href-no-hash'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/html-has-lang.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/html-has-lang'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/img-has-alt.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/img-has-alt'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/img-redundant-alt.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/img-redundant-alt'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/label-has-for.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/label-has-for'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/lang.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/lang'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/mouse-events-have-key-events.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/mouse-events-have-key-events'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/no-access-key.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/no-access-key'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/no-marquee.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/no-marquee'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/no-onchange.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/no-onchange'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/no-static-element-interactions.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/no-static-element-interactions'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/onclick-has-focus.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/onclick-has-focus'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/onclick-has-role.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/onclick-has-role'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/role-has-required-aria-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/role-has-required-aria-props'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/role-supports-aria-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/role-supports-aria-props'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/scope.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/scope'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/rules/tabindex-no-positive.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/rules/tabindex-no-positive'>;
-}
-declare module 'eslint-plugin-jsx-a11y/tests/src/util/getSuggestion.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/tests/src/util/getSuggestion'>;
+declare module 'eslint-plugin-jsx-a11y/src/util/schemas.js' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/src/util/schemas'>;
 }
