@@ -1,5 +1,5 @@
-// flow-typed signature: c69bf5bcc5825d5b6fbb9ef06b453857
-// flow-typed version: <<STUB>>/eslint_v^3.0.1/flow_v0.41.0
+// flow-typed signature: 9354e7847cf55f354b2ca85525e8513a
+// flow-typed version: <<STUB>>/eslint_v^3.19.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
