@@ -1,5 +1,5 @@
-// flow-typed signature: fda1031c122f7ca28baabb8a3e86d42f
-// flow-typed version: <<STUB>>/fetch-mock_v^5.5.0/flow_v0.41.0
+// flow-typed signature: 6601b56205e9dd4e89a30a1105cb846d
+// flow-typed version: <<STUB>>/fetch-mock_v^5.5.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
