@@ -1,5 +1,5 @@
-// flow-typed signature: 300a1c2ab51ec4d7895300a7c54db6fd
-// flow-typed version: <<STUB>>/flow-typed_v^2.0.0/flow_v0.41.0
+// flow-typed signature: 8898603ed592b58f9cb89fd30a371a09
+// flow-typed version: <<STUB>>/flow-typed_v^2.1.2/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
