@@ -1,5 +1,5 @@
-// flow-typed signature: 141c1b58d6681b9949e4465632de5c4e
-// flow-typed version: <<STUB>>/karma-phantomjs-launcher_v^1.0.2/flow_v0.41.0
+// flow-typed signature: 5d3ffe52b792255244d23d40db3a0eb5
+// flow-typed version: <<STUB>>/karma-phantomjs-launcher_v^1.0.4/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
