@@ -1,5 +1,5 @@
-// flow-typed signature: 0a8fdd93de5bb589ab38f4fc988c3add
-// flow-typed version: <<STUB>>/karma_v^1.0.0/flow_v0.41.0
+// flow-typed signature: 7f79ffd33957aa678d56a59fa0becebd
+// flow-typed version: <<STUB>>/karma_v^1.7.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
