@@ -1,5 +1,5 @@
-// flow-typed signature: 391dab989b00d2bb6b3063533c76680d
-// flow-typed version: <<STUB>>/node-sass_v^4.0.0/flow_v0.41.0
+// flow-typed signature: 1467897619c90e9b46bb6cd261c046b3
+// flow-typed version: <<STUB>>/node-sass_v^4.5.2/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
