@@ -1,5 +1,5 @@
-// flow-typed signature: af11c46fa20e5245f7ed2bf1e55713c3
-// flow-typed version: <<STUB>>/normalize.css_v^5.0.0/flow_v0.41.0
+// flow-typed signature: 71486ce2f5664b68bc233ddacb24c41b
+// flow-typed version: <<STUB>>/normalize.css_v^7.0.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
