@@ -1,5 +1,5 @@
-// flow-typed signature: 2dd757e34c0f09ede6e79211ac863b9e
-// flow-typed version: <<STUB>>/phantomjs-prebuilt_v^2.1.12/flow_v0.41.0
+// flow-typed signature: 07ffb6147c56622e0be0b340dda2db87
+// flow-typed version: <<STUB>>/phantomjs-prebuilt_v^2.1.14/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
