@@ -1,5 +1,5 @@
-// flow-typed signature: 77f34b700a61802cdf0a563c6e32fd8f
-// flow-typed version: <<STUB>>/react-bootstrap_v^0.30.7/flow_v0.41.0
+// flow-typed signature: 675542d4373c1b65ad5daccea2e823b7
+// flow-typed version: <<STUB>>/react-bootstrap_v^0.31.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -79,6 +79,10 @@ declare module 'react-bootstrap/es/Checkbox' {
 }
 
 declare module 'react-bootstrap/es/Clearfix' {
+  declare module.exports: any;
+}
+
+declare module 'react-bootstrap/es/CloseButton' {
   declare module.exports: any;
 }
 
@@ -470,6 +474,10 @@ declare module 'react-bootstrap/lib/Clearfix' {
   declare module.exports: any;
 }
 
+declare module 'react-bootstrap/lib/CloseButton' {
+  declare module.exports: any;
+}
+
 declare module 'react-bootstrap/lib/Col' {
   declare module.exports: any;
 }
@@ -852,6 +860,9 @@ declare module 'react-bootstrap/es/Checkbox.js' {
 declare module 'react-bootstrap/es/Clearfix.js' {
   declare module.exports: $Exports<'react-bootstrap/es/Clearfix'>;
 }
+declare module 'react-bootstrap/es/CloseButton.js' {
+  declare module.exports: $Exports<'react-bootstrap/es/CloseButton'>;
+}
 declare module 'react-bootstrap/es/Col.js' {
   declare module.exports: $Exports<'react-bootstrap/es/Col'>;
 }
@@ -1142,6 +1153,9 @@ declare module 'react-bootstrap/lib/Checkbox.js' {
 }
 declare module 'react-bootstrap/lib/Clearfix.js' {
   declare module.exports: $Exports<'react-bootstrap/lib/Clearfix'>;
+}
+declare module 'react-bootstrap/lib/CloseButton.js' {
+  declare module.exports: $Exports<'react-bootstrap/lib/CloseButton'>;
 }
 declare module 'react-bootstrap/lib/Col.js' {
   declare module.exports: $Exports<'react-bootstrap/lib/Col'>;
