@@ -1,5 +1,5 @@
-// flow-typed signature: 434e83948aef179af3cd7131161d7975
-// flow-typed version: <<STUB>>/react-icons_v^2.2.1/flow_v0.41.0
+// flow-typed signature: 3036d5365c6bf61b4d4185e4ee499e3b
+// flow-typed version: <<STUB>>/react-icons_v^2.2.5/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
