@@ -1,5 +1,5 @@
-// flow-typed signature: 23a697c680b6da680c36505dc20998ae
-// flow-typed version: <<STUB>>/react-s-alert_v^1.2.2/flow_v0.41.0
+// flow-typed signature: be0cbde6f733903a90c678c34d33c345
+// flow-typed version: <<STUB>>/react-s-alert_v^1.3.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
