@@ -1,5 +1,5 @@
-// flow-typed signature: 70ff50c6650c4c95a325aade98126702
-// flow-typed version: <<STUB>>/redux-actions_v^0.13.0/flow_v0.41.0
+// flow-typed signature: 2e86b051cdfc5545afacbc78e5450377
+// flow-typed version: <<STUB>>/redux-actions_v^2.0.3/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -30,6 +30,54 @@ declare module 'redux-actions/dist/redux-actions.min' {
   declare module.exports: any;
 }
 
+declare module 'redux-actions/es/arrayToObject' {
+  declare module.exports: any;
+}
+
+declare module 'redux-actions/es/camelCase' {
+  declare module.exports: any;
+}
+
+declare module 'redux-actions/es/combineActions' {
+  declare module.exports: any;
+}
+
+declare module 'redux-actions/es/createAction' {
+  declare module.exports: any;
+}
+
+declare module 'redux-actions/es/createActions' {
+  declare module.exports: any;
+}
+
+declare module 'redux-actions/es/handleAction' {
+  declare module.exports: any;
+}
+
+declare module 'redux-actions/es/handleActions' {
+  declare module.exports: any;
+}
+
+declare module 'redux-actions/es/index' {
+  declare module.exports: any;
+}
+
+declare module 'redux-actions/es/namespaceActions' {
+  declare module.exports: any;
+}
+
+declare module 'redux-actions/es/ownKeys' {
+  declare module.exports: any;
+}
+
+declare module 'redux-actions/lib/arrayToObject' {
+  declare module.exports: any;
+}
+
+declare module 'redux-actions/lib/camelCase' {
+  declare module.exports: any;
+}
+
 declare module 'redux-actions/lib/combineActions' {
   declare module.exports: any;
 }
@@ -54,6 +102,10 @@ declare module 'redux-actions/lib/index' {
   declare module.exports: any;
 }
 
+declare module 'redux-actions/lib/namespaceActions' {
+  declare module.exports: any;
+}
+
 declare module 'redux-actions/lib/ownKeys' {
   declare module.exports: any;
 }
@@ -64,6 +116,42 @@ declare module 'redux-actions/dist/redux-actions.js' {
 }
 declare module 'redux-actions/dist/redux-actions.min.js' {
   declare module.exports: $Exports<'redux-actions/dist/redux-actions.min'>;
+}
+declare module 'redux-actions/es/arrayToObject.js' {
+  declare module.exports: $Exports<'redux-actions/es/arrayToObject'>;
+}
+declare module 'redux-actions/es/camelCase.js' {
+  declare module.exports: $Exports<'redux-actions/es/camelCase'>;
+}
+declare module 'redux-actions/es/combineActions.js' {
+  declare module.exports: $Exports<'redux-actions/es/combineActions'>;
+}
+declare module 'redux-actions/es/createAction.js' {
+  declare module.exports: $Exports<'redux-actions/es/createAction'>;
+}
+declare module 'redux-actions/es/createActions.js' {
+  declare module.exports: $Exports<'redux-actions/es/createActions'>;
+}
+declare module 'redux-actions/es/handleAction.js' {
+  declare module.exports: $Exports<'redux-actions/es/handleAction'>;
+}
+declare module 'redux-actions/es/handleActions.js' {
+  declare module.exports: $Exports<'redux-actions/es/handleActions'>;
+}
+declare module 'redux-actions/es/index.js' {
+  declare module.exports: $Exports<'redux-actions/es/index'>;
+}
+declare module 'redux-actions/es/namespaceActions.js' {
+  declare module.exports: $Exports<'redux-actions/es/namespaceActions'>;
+}
+declare module 'redux-actions/es/ownKeys.js' {
+  declare module.exports: $Exports<'redux-actions/es/ownKeys'>;
+}
+declare module 'redux-actions/lib/arrayToObject.js' {
+  declare module.exports: $Exports<'redux-actions/lib/arrayToObject'>;
+}
+declare module 'redux-actions/lib/camelCase.js' {
+  declare module.exports: $Exports<'redux-actions/lib/camelCase'>;
 }
 declare module 'redux-actions/lib/combineActions.js' {
   declare module.exports: $Exports<'redux-actions/lib/combineActions'>;
@@ -82,6 +170,9 @@ declare module 'redux-actions/lib/handleActions.js' {
 }
 declare module 'redux-actions/lib/index.js' {
   declare module.exports: $Exports<'redux-actions/lib/index'>;
+}
+declare module 'redux-actions/lib/namespaceActions.js' {
+  declare module.exports: $Exports<'redux-actions/lib/namespaceActions'>;
 }
 declare module 'redux-actions/lib/ownKeys.js' {
   declare module.exports: $Exports<'redux-actions/lib/ownKeys'>;
