@@ -1,5 +1,5 @@
-// flow-typed signature: c6cf5f6ce0288a9d2c6e8fdd6345ebc7
-// flow-typed version: <<STUB>>/redux-auth-wrapper_v^1.0.0/flow_v0.41.0
+// flow-typed signature: 8d57a901aa15663254de1722ae977c87
+// flow-typed version: <<STUB>>/redux-auth-wrapper_v^1.1.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
