@@ -1,5 +1,5 @@
-// flow-typed signature: e7a1fddddd2d807819ac92f974d534b0
-// flow-typed version: <<STUB>>/redux-saga-test-plan_v^3.0.1/flow_v0.41.0
+// flow-typed signature: a80aae2477b9553814eca30a5aef0d13
+// flow-typed version: <<STUB>>/redux-saga-test-plan_v^3.0.2/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
