@@ -1,5 +1,5 @@
-// flow-typed signature: ca7fb63d4fef6ec70fd0a7034e54ff5a
-// flow-typed version: <<STUB>>/redux-saga_v^0.15.1/flow_v0.41.0
+// flow-typed signature: 7e6a1dd852d8c467b423b70e0d946e2c
+// flow-typed version: <<STUB>>/redux-saga_v^0.15.3/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
