@@ -1,5 +1,5 @@
-// flow-typed signature: 6ca26e7feb6ec8c06e70a79c83bd93a7
-// flow-typed version: <<STUB>>/redux-thunk_v^2.0.0/flow_v0.41.0
+// flow-typed signature: 944a1150d22721dc5aeb940c62aa9231
+// flow-typed version: <<STUB>>/redux-thunk_v^2.2.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
