@@ -1,5 +1,5 @@
-// flow-typed signature: ce93c2e8b7a44f17cc5d5f213c9f2d18
-// flow-typed version: <<STUB>>/rimraf_v^2.5.4/flow_v0.41.0
+// flow-typed signature: 475d59277bfba3e04a313defb589ae39
+// flow-typed version: <<STUB>>/rimraf_v^2.6.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
