@@ -1,5 +1,5 @@
-// flow-typed signature: 6f936df8033e925e9d55b5f371665ea4
-// flow-typed version: <<STUB>>/sass-loader_v^4.0.0/flow_v0.41.0
+// flow-typed signature: 6efcdbf36291d99e985beb36f95158fd
+// flow-typed version: <<STUB>>/sass-loader_v^6.0.3/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,12 +22,46 @@ declare module 'sass-loader' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
+declare module 'sass-loader/lib/formatSassError' {
+  declare module.exports: any;
+}
 
+declare module 'sass-loader/lib/importsToResolve' {
+  declare module.exports: any;
+}
+
+declare module 'sass-loader/lib/loader' {
+  declare module.exports: any;
+}
+
+declare module 'sass-loader/lib/normalizeOptions' {
+  declare module.exports: any;
+}
+
+declare module 'sass-loader/lib/proxyCustomImporters' {
+  declare module.exports: any;
+}
+
+declare module 'sass-loader/lib/webpackImporter' {
+  declare module.exports: any;
+}
 
 // Filename aliases
-declare module 'sass-loader/index' {
-  declare module.exports: $Exports<'sass-loader'>;
+declare module 'sass-loader/lib/formatSassError.js' {
+  declare module.exports: $Exports<'sass-loader/lib/formatSassError'>;
 }
-declare module 'sass-loader/index.js' {
-  declare module.exports: $Exports<'sass-loader'>;
+declare module 'sass-loader/lib/importsToResolve.js' {
+  declare module.exports: $Exports<'sass-loader/lib/importsToResolve'>;
+}
+declare module 'sass-loader/lib/loader.js' {
+  declare module.exports: $Exports<'sass-loader/lib/loader'>;
+}
+declare module 'sass-loader/lib/normalizeOptions.js' {
+  declare module.exports: $Exports<'sass-loader/lib/normalizeOptions'>;
+}
+declare module 'sass-loader/lib/proxyCustomImporters.js' {
+  declare module.exports: $Exports<'sass-loader/lib/proxyCustomImporters'>;
+}
+declare module 'sass-loader/lib/webpackImporter.js' {
+  declare module.exports: $Exports<'sass-loader/lib/webpackImporter'>;
 }
