@@ -1,5 +1,5 @@
-// flow-typed signature: f15e68eae834fd281877efd8f959a1a5
-// flow-typed version: <<STUB>>/sinon-chai_v^2.8.0/flow_v0.41.0
+// flow-typed signature: 6fd493009bd1228942a05a0234b031b1
+// flow-typed version: <<STUB>>/sinon-chai_v^2.8.0/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
