@@ -1,5 +1,5 @@
-// flow-typed signature: bf6bf491cea1004ae0acea8468248a2e
-// flow-typed version: <<STUB>>/sinon_v^1.17.5/flow_v0.41.0
+// flow-typed signature: 532aca95f7f0df52ea26565c7106cffe
+// flow-typed version: <<STUB>>/sinon_v^1.17.5/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
