@@ -1,5 +1,5 @@
-// flow-typed signature: ac2f843d4e012759f7a8f75bfeb6fc71
-// flow-typed version: <<STUB>>/style-loader_v^0.13.1/flow_v0.41.0
+// flow-typed signature: 53d7f163417d3c479ddaaa89d3adc18d
+// flow-typed version: <<STUB>>/style-loader_v^0.16.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -30,6 +30,22 @@ declare module 'style-loader/addStyleUrl' {
   declare module.exports: any;
 }
 
+declare module 'style-loader/fixUrls' {
+  declare module.exports: any;
+}
+
+declare module 'style-loader/test/basicTest' {
+  declare module.exports: any;
+}
+
+declare module 'style-loader/test/fixUrlsTest' {
+  declare module.exports: any;
+}
+
+declare module 'style-loader/test/utils' {
+  declare module.exports: any;
+}
+
 declare module 'style-loader/url' {
   declare module.exports: any;
 }
@@ -45,11 +61,23 @@ declare module 'style-loader/addStyles.js' {
 declare module 'style-loader/addStyleUrl.js' {
   declare module.exports: $Exports<'style-loader/addStyleUrl'>;
 }
+declare module 'style-loader/fixUrls.js' {
+  declare module.exports: $Exports<'style-loader/fixUrls'>;
+}
 declare module 'style-loader/index' {
   declare module.exports: $Exports<'style-loader'>;
 }
 declare module 'style-loader/index.js' {
   declare module.exports: $Exports<'style-loader'>;
+}
+declare module 'style-loader/test/basicTest.js' {
+  declare module.exports: $Exports<'style-loader/test/basicTest'>;
+}
+declare module 'style-loader/test/fixUrlsTest.js' {
+  declare module.exports: $Exports<'style-loader/test/fixUrlsTest'>;
+}
+declare module 'style-loader/test/utils.js' {
+  declare module.exports: $Exports<'style-loader/test/utils'>;
 }
 declare module 'style-loader/url.js' {
   declare module.exports: $Exports<'style-loader/url'>;
