@@ -1,5 +1,5 @@
-// flow-typed signature: 301ce11a2b7b30795a74ae4856eb3297
-// flow-typed version: <<STUB>>/webpack_v^1.12.14/flow_v0.41.0
+// flow-typed signature: c8b66fcbc112807e27438f7bd99c77b4
+// flow-typed version: <<STUB>>/webpack_v^2.5.1/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -26,6 +26,10 @@ declare module 'webpack/bin/config-optimist' {
   declare module.exports: any;
 }
 
+declare module 'webpack/bin/config-yargs' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/bin/convert-argv' {
   declare module.exports: any;
 }
@@ -42,15 +46,27 @@ declare module 'webpack/buildin/amd-options' {
   declare module.exports: any;
 }
 
+declare module 'webpack/buildin/global' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/buildin/harmony-module' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/buildin/module' {
   declare module.exports: any;
 }
 
-declare module 'webpack/buildin/return-require' {
+declare module 'webpack/buildin/system' {
   declare module.exports: any;
 }
 
 declare module 'webpack/hot/dev-server' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/hot/emitter' {
   declare module.exports: any;
 }
 
@@ -70,19 +86,11 @@ declare module 'webpack/hot/signal' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/AbstractPlugin' {
-  declare module.exports: any;
-}
-
 declare module 'webpack/lib/AmdMainTemplatePlugin' {
   declare module.exports: any;
 }
 
 declare module 'webpack/lib/APIPlugin' {
-  declare module.exports: any;
-}
-
-declare module 'webpack/lib/ArrayMap' {
   declare module.exports: any;
 }
 
@@ -122,6 +130,10 @@ declare module 'webpack/lib/ChunkTemplate' {
   declare module.exports: any;
 }
 
+declare module 'webpack/lib/compareLocations' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/lib/CompatibilityPlugin' {
   declare module.exports: any;
 }
@@ -131,10 +143,6 @@ declare module 'webpack/lib/Compilation' {
 }
 
 declare module 'webpack/lib/Compiler' {
-  declare module.exports: any;
-}
-
-declare module 'webpack/lib/ConcatSource' {
   declare module.exports: any;
 }
 
@@ -151,10 +159,6 @@ declare module 'webpack/lib/ContextModuleFactory' {
 }
 
 declare module 'webpack/lib/ContextReplacementPlugin' {
-  declare module.exports: any;
-}
-
-declare module 'webpack/lib/CriticalDependenciesWarning' {
   declare module.exports: any;
 }
 
@@ -250,6 +254,10 @@ declare module 'webpack/lib/dependencies/ContextElementDependency' {
   declare module.exports: any;
 }
 
+declare module 'webpack/lib/dependencies/CriticalDependencyWarning' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/lib/dependencies/DelegatedSourceDependency' {
   declare module.exports: any;
 }
@@ -266,19 +274,79 @@ declare module 'webpack/lib/dependencies/getFunctionExpression' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/dependencies/LabeledExportsDependency' {
+declare module 'webpack/lib/dependencies/HarmonyAcceptDependency' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/dependencies/LabeledModuleDependency' {
+declare module 'webpack/lib/dependencies/HarmonyAcceptImportDependency' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/dependencies/LabeledModuleDependencyParserPlugin' {
+declare module 'webpack/lib/dependencies/HarmonyCompatibilityDependency' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/dependencies/LabeledModulesPlugin' {
+declare module 'webpack/lib/dependencies/HarmonyDetectionParserPlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/HarmonyExportDependencyParserPlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/HarmonyExportExpressionDependency' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/HarmonyExportHeaderDependency' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/HarmonyExportImportedSpecifierDependency' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/HarmonyExportSpecifierDependency' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/HarmonyImportDependency' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/HarmonyImportDependencyParserPlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/HarmonyImportSpecifierDependency' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/HarmonyModulesHelpers' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/HarmonyModulesPlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/ImportContextDependency' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/ImportDependenciesBlock' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/ImportDependency' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/ImportParserPlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/ImportPlugin' {
   declare module.exports: any;
 }
 
@@ -327,10 +395,6 @@ declare module 'webpack/lib/dependencies/MultiEntryDependency' {
 }
 
 declare module 'webpack/lib/dependencies/NullDependency' {
-  declare module.exports: any;
-}
-
-declare module 'webpack/lib/dependencies/NullDependencyTemplate' {
   declare module.exports: any;
 }
 
@@ -406,7 +470,11 @@ declare module 'webpack/lib/dependencies/SingleEntryDependency' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/dependencies/TemplateArgumentDependency' {
+declare module 'webpack/lib/dependencies/SystemPlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/dependencies/UnsupportedDependency' {
   declare module.exports: any;
 }
 
@@ -446,6 +514,10 @@ declare module 'webpack/lib/DllReferencePlugin' {
   declare module.exports: any;
 }
 
+declare module 'webpack/lib/DynamicEntryPlugin' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/lib/EntryModuleNotFoundError' {
   declare module.exports: any;
 }
@@ -454,7 +526,15 @@ declare module 'webpack/lib/EntryOptionPlugin' {
   declare module.exports: any;
 }
 
+declare module 'webpack/lib/Entrypoint' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/lib/EnvironmentPlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/ErrorHelpers' {
   declare module.exports: any;
 }
 
@@ -490,11 +570,31 @@ declare module 'webpack/lib/ExternalsPlugin' {
   declare module.exports: any;
 }
 
+declare module 'webpack/lib/FlagDependencyExportsPlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/FlagDependencyUsagePlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/FlagInitialModulesAsUsedPlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/formatLocation' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/lib/FunctionModulePlugin' {
   declare module.exports: any;
 }
 
 declare module 'webpack/lib/FunctionModuleTemplatePlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/HashedModuleIdsPlugin' {
   declare module.exports: any;
 }
 
@@ -546,6 +646,10 @@ declare module 'webpack/lib/LibraryTemplatePlugin' {
   declare module.exports: any;
 }
 
+declare module 'webpack/lib/LoaderOptionsPlugin' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/lib/LoaderTargetPlugin' {
   declare module.exports: any;
 }
@@ -562,6 +666,22 @@ declare module 'webpack/lib/Module' {
   declare module.exports: any;
 }
 
+declare module 'webpack/lib/ModuleBuildError' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/ModuleDependencyError' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/ModuleDependencyWarning' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/ModuleError' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/lib/ModuleFilenameHelpers' {
   declare module.exports: any;
 }
@@ -574,15 +694,15 @@ declare module 'webpack/lib/ModuleParseError' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/ModuleParserHelpers' {
-  declare module.exports: any;
-}
-
 declare module 'webpack/lib/ModuleReason' {
   declare module.exports: any;
 }
 
 declare module 'webpack/lib/ModuleTemplate' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/ModuleWarning' {
   declare module.exports: any;
 }
 
@@ -603,6 +723,18 @@ declare module 'webpack/lib/MultiModule' {
 }
 
 declare module 'webpack/lib/MultiModuleFactory' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/MultiStats' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/MultiWatching' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/NamedChunksPlugin' {
   declare module.exports: any;
 }
 
@@ -658,11 +790,11 @@ declare module 'webpack/lib/node/NodeWatchFileSystem' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/node/OldNodeWatchFileSystem' {
+declare module 'webpack/lib/NodeStuffPlugin' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/NodeStuffPlugin' {
+declare module 'webpack/lib/NoEmitOnErrorsPlugin' {
   declare module.exports: any;
 }
 
@@ -686,11 +818,15 @@ declare module 'webpack/lib/NullFactory' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/OldWatchingPlugin' {
+declare module 'webpack/lib/optimize/AggressiveMergingPlugin' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/optimize/AggressiveMergingPlugin' {
+declare module 'webpack/lib/optimize/AggressiveSplittingPlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/optimize/ChunkModuleIdRangePlugin' {
   declare module.exports: any;
 }
 
@@ -699,6 +835,10 @@ declare module 'webpack/lib/optimize/CommonsChunkPlugin' {
 }
 
 declare module 'webpack/lib/optimize/DedupePlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/optimize/EnsureChunkConditionsPlugin' {
   declare module.exports: any;
 }
 
@@ -715,10 +855,6 @@ declare module 'webpack/lib/optimize/MergeDuplicateChunksPlugin' {
 }
 
 declare module 'webpack/lib/optimize/MinChunkSizePlugin' {
-  declare module.exports: any;
-}
-
-declare module 'webpack/lib/optimize/OccurenceOrderPlugin' {
   declare module.exports: any;
 }
 
@@ -742,11 +878,31 @@ declare module 'webpack/lib/OptionsApply' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/OriginalSource' {
+declare module 'webpack/lib/OptionsDefaulter' {
   declare module.exports: any;
 }
 
 declare module 'webpack/lib/Parser' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/ParserHelpers' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/performance/AssetsOverSizeLimitWarning' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/performance/EntrypointsOverSizeLimitWarning' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/performance/NoAsyncChunksWarning' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/performance/SizeLimitsPlugin' {
   declare module.exports: any;
 }
 
@@ -766,10 +922,6 @@ declare module 'webpack/lib/RawModule' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/RawSource' {
-  declare module.exports: any;
-}
-
 declare module 'webpack/lib/RecordIdsPlugin' {
   declare module.exports: any;
 }
@@ -786,7 +938,7 @@ declare module 'webpack/lib/RequireJsStuffPlugin' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/ResolverPlugin' {
+declare module 'webpack/lib/RuleSet' {
   declare module.exports: any;
 }
 
@@ -798,7 +950,7 @@ declare module 'webpack/lib/SingleEntryPlugin' {
   declare module.exports: any;
 }
 
-declare module 'webpack/lib/Source' {
+declare module 'webpack/lib/SizeFormatHelpers' {
   declare module.exports: any;
 }
 
@@ -807,10 +959,6 @@ declare module 'webpack/lib/SourceMapDevToolModuleOptionsPlugin' {
 }
 
 declare module 'webpack/lib/SourceMapDevToolPlugin' {
-  declare module.exports: any;
-}
-
-declare module 'webpack/lib/SourceMapSource' {
   declare module.exports: any;
 }
 
@@ -834,6 +982,18 @@ declare module 'webpack/lib/UnsupportedFeatureWarning' {
   declare module.exports: any;
 }
 
+declare module 'webpack/lib/UseStrictPlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/util/identifier' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/validateSchema' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/lib/WarnCaseSensitiveModulesPlugin' {
   declare module.exports: any;
 }
@@ -854,6 +1014,10 @@ declare module 'webpack/lib/webpack.web' {
   declare module.exports: any;
 }
 
+declare module 'webpack/lib/WebpackError' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/lib/WebpackOptionsApply' {
   declare module.exports: any;
 }
@@ -862,7 +1026,19 @@ declare module 'webpack/lib/WebpackOptionsDefaulter' {
   declare module.exports: any;
 }
 
+declare module 'webpack/lib/WebpackOptionsValidationError' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/lib/webworker/WebWorkerChunkTemplatePlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/webworker/WebWorkerHotUpdateChunkTemplatePlugin' {
+  declare module.exports: any;
+}
+
+declare module 'webpack/lib/webworker/WebWorkerMainTemplate.runtime' {
   declare module.exports: any;
 }
 
@@ -874,6 +1050,10 @@ declare module 'webpack/lib/webworker/WebWorkerTemplatePlugin' {
   declare module.exports: any;
 }
 
+declare module 'webpack/schemas/ajv.absolutePath' {
+  declare module.exports: any;
+}
+
 declare module 'webpack/web_modules/node-libs-browser' {
   declare module.exports: any;
 }
@@ -881,6 +1061,9 @@ declare module 'webpack/web_modules/node-libs-browser' {
 // Filename aliases
 declare module 'webpack/bin/config-optimist.js' {
   declare module.exports: $Exports<'webpack/bin/config-optimist'>;
+}
+declare module 'webpack/bin/config-yargs.js' {
+  declare module.exports: $Exports<'webpack/bin/config-yargs'>;
 }
 declare module 'webpack/bin/convert-argv.js' {
   declare module.exports: $Exports<'webpack/bin/convert-argv'>;
@@ -894,14 +1077,23 @@ declare module 'webpack/buildin/amd-define.js' {
 declare module 'webpack/buildin/amd-options.js' {
   declare module.exports: $Exports<'webpack/buildin/amd-options'>;
 }
+declare module 'webpack/buildin/global.js' {
+  declare module.exports: $Exports<'webpack/buildin/global'>;
+}
+declare module 'webpack/buildin/harmony-module.js' {
+  declare module.exports: $Exports<'webpack/buildin/harmony-module'>;
+}
 declare module 'webpack/buildin/module.js' {
   declare module.exports: $Exports<'webpack/buildin/module'>;
 }
-declare module 'webpack/buildin/return-require.js' {
-  declare module.exports: $Exports<'webpack/buildin/return-require'>;
+declare module 'webpack/buildin/system.js' {
+  declare module.exports: $Exports<'webpack/buildin/system'>;
 }
 declare module 'webpack/hot/dev-server.js' {
   declare module.exports: $Exports<'webpack/hot/dev-server'>;
+}
+declare module 'webpack/hot/emitter.js' {
+  declare module.exports: $Exports<'webpack/hot/emitter'>;
 }
 declare module 'webpack/hot/log-apply-result.js' {
   declare module.exports: $Exports<'webpack/hot/log-apply-result'>;
@@ -915,17 +1107,11 @@ declare module 'webpack/hot/poll.js' {
 declare module 'webpack/hot/signal.js' {
   declare module.exports: $Exports<'webpack/hot/signal'>;
 }
-declare module 'webpack/lib/AbstractPlugin.js' {
-  declare module.exports: $Exports<'webpack/lib/AbstractPlugin'>;
-}
 declare module 'webpack/lib/AmdMainTemplatePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/AmdMainTemplatePlugin'>;
 }
 declare module 'webpack/lib/APIPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/APIPlugin'>;
-}
-declare module 'webpack/lib/ArrayMap.js' {
-  declare module.exports: $Exports<'webpack/lib/ArrayMap'>;
 }
 declare module 'webpack/lib/AsyncDependenciesBlock.js' {
   declare module.exports: $Exports<'webpack/lib/AsyncDependenciesBlock'>;
@@ -954,6 +1140,9 @@ declare module 'webpack/lib/ChunkRenderError.js' {
 declare module 'webpack/lib/ChunkTemplate.js' {
   declare module.exports: $Exports<'webpack/lib/ChunkTemplate'>;
 }
+declare module 'webpack/lib/compareLocations.js' {
+  declare module.exports: $Exports<'webpack/lib/compareLocations'>;
+}
 declare module 'webpack/lib/CompatibilityPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/CompatibilityPlugin'>;
 }
@@ -962,9 +1151,6 @@ declare module 'webpack/lib/Compilation.js' {
 }
 declare module 'webpack/lib/Compiler.js' {
   declare module.exports: $Exports<'webpack/lib/Compiler'>;
-}
-declare module 'webpack/lib/ConcatSource.js' {
-  declare module.exports: $Exports<'webpack/lib/ConcatSource'>;
 }
 declare module 'webpack/lib/ConstPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/ConstPlugin'>;
@@ -977,9 +1163,6 @@ declare module 'webpack/lib/ContextModuleFactory.js' {
 }
 declare module 'webpack/lib/ContextReplacementPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/ContextReplacementPlugin'>;
-}
-declare module 'webpack/lib/CriticalDependenciesWarning.js' {
-  declare module.exports: $Exports<'webpack/lib/CriticalDependenciesWarning'>;
 }
 declare module 'webpack/lib/DefinePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/DefinePlugin'>;
@@ -1050,6 +1233,9 @@ declare module 'webpack/lib/dependencies/ContextDependencyTemplateAsRequireCall.
 declare module 'webpack/lib/dependencies/ContextElementDependency.js' {
   declare module.exports: $Exports<'webpack/lib/dependencies/ContextElementDependency'>;
 }
+declare module 'webpack/lib/dependencies/CriticalDependencyWarning.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/CriticalDependencyWarning'>;
+}
 declare module 'webpack/lib/dependencies/DelegatedSourceDependency.js' {
   declare module.exports: $Exports<'webpack/lib/dependencies/DelegatedSourceDependency'>;
 }
@@ -1062,17 +1248,62 @@ declare module 'webpack/lib/dependencies/DllEntryDependency.js' {
 declare module 'webpack/lib/dependencies/getFunctionExpression.js' {
   declare module.exports: $Exports<'webpack/lib/dependencies/getFunctionExpression'>;
 }
-declare module 'webpack/lib/dependencies/LabeledExportsDependency.js' {
-  declare module.exports: $Exports<'webpack/lib/dependencies/LabeledExportsDependency'>;
+declare module 'webpack/lib/dependencies/HarmonyAcceptDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyAcceptDependency'>;
 }
-declare module 'webpack/lib/dependencies/LabeledModuleDependency.js' {
-  declare module.exports: $Exports<'webpack/lib/dependencies/LabeledModuleDependency'>;
+declare module 'webpack/lib/dependencies/HarmonyAcceptImportDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyAcceptImportDependency'>;
 }
-declare module 'webpack/lib/dependencies/LabeledModuleDependencyParserPlugin.js' {
-  declare module.exports: $Exports<'webpack/lib/dependencies/LabeledModuleDependencyParserPlugin'>;
+declare module 'webpack/lib/dependencies/HarmonyCompatibilityDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyCompatibilityDependency'>;
 }
-declare module 'webpack/lib/dependencies/LabeledModulesPlugin.js' {
-  declare module.exports: $Exports<'webpack/lib/dependencies/LabeledModulesPlugin'>;
+declare module 'webpack/lib/dependencies/HarmonyDetectionParserPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyDetectionParserPlugin'>;
+}
+declare module 'webpack/lib/dependencies/HarmonyExportDependencyParserPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyExportDependencyParserPlugin'>;
+}
+declare module 'webpack/lib/dependencies/HarmonyExportExpressionDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyExportExpressionDependency'>;
+}
+declare module 'webpack/lib/dependencies/HarmonyExportHeaderDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyExportHeaderDependency'>;
+}
+declare module 'webpack/lib/dependencies/HarmonyExportImportedSpecifierDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyExportImportedSpecifierDependency'>;
+}
+declare module 'webpack/lib/dependencies/HarmonyExportSpecifierDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyExportSpecifierDependency'>;
+}
+declare module 'webpack/lib/dependencies/HarmonyImportDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyImportDependency'>;
+}
+declare module 'webpack/lib/dependencies/HarmonyImportDependencyParserPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyImportDependencyParserPlugin'>;
+}
+declare module 'webpack/lib/dependencies/HarmonyImportSpecifierDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyImportSpecifierDependency'>;
+}
+declare module 'webpack/lib/dependencies/HarmonyModulesHelpers.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyModulesHelpers'>;
+}
+declare module 'webpack/lib/dependencies/HarmonyModulesPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/HarmonyModulesPlugin'>;
+}
+declare module 'webpack/lib/dependencies/ImportContextDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/ImportContextDependency'>;
+}
+declare module 'webpack/lib/dependencies/ImportDependenciesBlock.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/ImportDependenciesBlock'>;
+}
+declare module 'webpack/lib/dependencies/ImportDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/ImportDependency'>;
+}
+declare module 'webpack/lib/dependencies/ImportParserPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/ImportParserPlugin'>;
+}
+declare module 'webpack/lib/dependencies/ImportPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/ImportPlugin'>;
 }
 declare module 'webpack/lib/dependencies/LoaderDependency.js' {
   declare module.exports: $Exports<'webpack/lib/dependencies/LoaderDependency'>;
@@ -1109,9 +1340,6 @@ declare module 'webpack/lib/dependencies/MultiEntryDependency.js' {
 }
 declare module 'webpack/lib/dependencies/NullDependency.js' {
   declare module.exports: $Exports<'webpack/lib/dependencies/NullDependency'>;
-}
-declare module 'webpack/lib/dependencies/NullDependencyTemplate.js' {
-  declare module.exports: $Exports<'webpack/lib/dependencies/NullDependencyTemplate'>;
 }
 declare module 'webpack/lib/dependencies/PrefetchDependency.js' {
   declare module.exports: $Exports<'webpack/lib/dependencies/PrefetchDependency'>;
@@ -1167,8 +1395,11 @@ declare module 'webpack/lib/dependencies/RequireResolveHeaderDependency.js' {
 declare module 'webpack/lib/dependencies/SingleEntryDependency.js' {
   declare module.exports: $Exports<'webpack/lib/dependencies/SingleEntryDependency'>;
 }
-declare module 'webpack/lib/dependencies/TemplateArgumentDependency.js' {
-  declare module.exports: $Exports<'webpack/lib/dependencies/TemplateArgumentDependency'>;
+declare module 'webpack/lib/dependencies/SystemPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/SystemPlugin'>;
+}
+declare module 'webpack/lib/dependencies/UnsupportedDependency.js' {
+  declare module.exports: $Exports<'webpack/lib/dependencies/UnsupportedDependency'>;
 }
 declare module 'webpack/lib/dependencies/WebpackMissingModule.js' {
   declare module.exports: $Exports<'webpack/lib/dependencies/WebpackMissingModule'>;
@@ -1197,14 +1428,23 @@ declare module 'webpack/lib/DllPlugin.js' {
 declare module 'webpack/lib/DllReferencePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/DllReferencePlugin'>;
 }
+declare module 'webpack/lib/DynamicEntryPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/DynamicEntryPlugin'>;
+}
 declare module 'webpack/lib/EntryModuleNotFoundError.js' {
   declare module.exports: $Exports<'webpack/lib/EntryModuleNotFoundError'>;
 }
 declare module 'webpack/lib/EntryOptionPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/EntryOptionPlugin'>;
 }
+declare module 'webpack/lib/Entrypoint.js' {
+  declare module.exports: $Exports<'webpack/lib/Entrypoint'>;
+}
 declare module 'webpack/lib/EnvironmentPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/EnvironmentPlugin'>;
+}
+declare module 'webpack/lib/ErrorHelpers.js' {
+  declare module.exports: $Exports<'webpack/lib/ErrorHelpers'>;
 }
 declare module 'webpack/lib/EvalDevToolModulePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/EvalDevToolModulePlugin'>;
@@ -1230,11 +1470,26 @@ declare module 'webpack/lib/ExternalModuleFactoryPlugin.js' {
 declare module 'webpack/lib/ExternalsPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/ExternalsPlugin'>;
 }
+declare module 'webpack/lib/FlagDependencyExportsPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/FlagDependencyExportsPlugin'>;
+}
+declare module 'webpack/lib/FlagDependencyUsagePlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/FlagDependencyUsagePlugin'>;
+}
+declare module 'webpack/lib/FlagInitialModulesAsUsedPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/FlagInitialModulesAsUsedPlugin'>;
+}
+declare module 'webpack/lib/formatLocation.js' {
+  declare module.exports: $Exports<'webpack/lib/formatLocation'>;
+}
 declare module 'webpack/lib/FunctionModulePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/FunctionModulePlugin'>;
 }
 declare module 'webpack/lib/FunctionModuleTemplatePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/FunctionModuleTemplatePlugin'>;
+}
+declare module 'webpack/lib/HashedModuleIdsPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/HashedModuleIdsPlugin'>;
 }
 declare module 'webpack/lib/HotModuleReplacement.runtime.js' {
   declare module.exports: $Exports<'webpack/lib/HotModuleReplacement.runtime'>;
@@ -1272,6 +1527,9 @@ declare module 'webpack/lib/LibManifestPlugin.js' {
 declare module 'webpack/lib/LibraryTemplatePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/LibraryTemplatePlugin'>;
 }
+declare module 'webpack/lib/LoaderOptionsPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/LoaderOptionsPlugin'>;
+}
 declare module 'webpack/lib/LoaderTargetPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/LoaderTargetPlugin'>;
 }
@@ -1284,6 +1542,18 @@ declare module 'webpack/lib/MemoryOutputFileSystem.js' {
 declare module 'webpack/lib/Module.js' {
   declare module.exports: $Exports<'webpack/lib/Module'>;
 }
+declare module 'webpack/lib/ModuleBuildError.js' {
+  declare module.exports: $Exports<'webpack/lib/ModuleBuildError'>;
+}
+declare module 'webpack/lib/ModuleDependencyError.js' {
+  declare module.exports: $Exports<'webpack/lib/ModuleDependencyError'>;
+}
+declare module 'webpack/lib/ModuleDependencyWarning.js' {
+  declare module.exports: $Exports<'webpack/lib/ModuleDependencyWarning'>;
+}
+declare module 'webpack/lib/ModuleError.js' {
+  declare module.exports: $Exports<'webpack/lib/ModuleError'>;
+}
 declare module 'webpack/lib/ModuleFilenameHelpers.js' {
   declare module.exports: $Exports<'webpack/lib/ModuleFilenameHelpers'>;
 }
@@ -1293,14 +1563,14 @@ declare module 'webpack/lib/ModuleNotFoundError.js' {
 declare module 'webpack/lib/ModuleParseError.js' {
   declare module.exports: $Exports<'webpack/lib/ModuleParseError'>;
 }
-declare module 'webpack/lib/ModuleParserHelpers.js' {
-  declare module.exports: $Exports<'webpack/lib/ModuleParserHelpers'>;
-}
 declare module 'webpack/lib/ModuleReason.js' {
   declare module.exports: $Exports<'webpack/lib/ModuleReason'>;
 }
 declare module 'webpack/lib/ModuleTemplate.js' {
   declare module.exports: $Exports<'webpack/lib/ModuleTemplate'>;
+}
+declare module 'webpack/lib/ModuleWarning.js' {
+  declare module.exports: $Exports<'webpack/lib/ModuleWarning'>;
 }
 declare module 'webpack/lib/MovedToPluginWarningPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/MovedToPluginWarningPlugin'>;
@@ -1316,6 +1586,15 @@ declare module 'webpack/lib/MultiModule.js' {
 }
 declare module 'webpack/lib/MultiModuleFactory.js' {
   declare module.exports: $Exports<'webpack/lib/MultiModuleFactory'>;
+}
+declare module 'webpack/lib/MultiStats.js' {
+  declare module.exports: $Exports<'webpack/lib/MultiStats'>;
+}
+declare module 'webpack/lib/MultiWatching.js' {
+  declare module.exports: $Exports<'webpack/lib/MultiWatching'>;
+}
+declare module 'webpack/lib/NamedChunksPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/NamedChunksPlugin'>;
 }
 declare module 'webpack/lib/NamedModulesPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/NamedModulesPlugin'>;
@@ -1356,11 +1635,11 @@ declare module 'webpack/lib/node/NodeTemplatePlugin.js' {
 declare module 'webpack/lib/node/NodeWatchFileSystem.js' {
   declare module.exports: $Exports<'webpack/lib/node/NodeWatchFileSystem'>;
 }
-declare module 'webpack/lib/node/OldNodeWatchFileSystem.js' {
-  declare module.exports: $Exports<'webpack/lib/node/OldNodeWatchFileSystem'>;
-}
 declare module 'webpack/lib/NodeStuffPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/NodeStuffPlugin'>;
+}
+declare module 'webpack/lib/NoEmitOnErrorsPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/NoEmitOnErrorsPlugin'>;
 }
 declare module 'webpack/lib/NoErrorsPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/NoErrorsPlugin'>;
@@ -1377,17 +1656,23 @@ declare module 'webpack/lib/NormalModuleReplacementPlugin.js' {
 declare module 'webpack/lib/NullFactory.js' {
   declare module.exports: $Exports<'webpack/lib/NullFactory'>;
 }
-declare module 'webpack/lib/OldWatchingPlugin.js' {
-  declare module.exports: $Exports<'webpack/lib/OldWatchingPlugin'>;
-}
 declare module 'webpack/lib/optimize/AggressiveMergingPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/optimize/AggressiveMergingPlugin'>;
+}
+declare module 'webpack/lib/optimize/AggressiveSplittingPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/optimize/AggressiveSplittingPlugin'>;
+}
+declare module 'webpack/lib/optimize/ChunkModuleIdRangePlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/optimize/ChunkModuleIdRangePlugin'>;
 }
 declare module 'webpack/lib/optimize/CommonsChunkPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/optimize/CommonsChunkPlugin'>;
 }
 declare module 'webpack/lib/optimize/DedupePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/optimize/DedupePlugin'>;
+}
+declare module 'webpack/lib/optimize/EnsureChunkConditionsPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/optimize/EnsureChunkConditionsPlugin'>;
 }
 declare module 'webpack/lib/optimize/FlagIncludedChunksPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/optimize/FlagIncludedChunksPlugin'>;
@@ -1400,9 +1685,6 @@ declare module 'webpack/lib/optimize/MergeDuplicateChunksPlugin.js' {
 }
 declare module 'webpack/lib/optimize/MinChunkSizePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/optimize/MinChunkSizePlugin'>;
-}
-declare module 'webpack/lib/optimize/OccurenceOrderPlugin.js' {
-  declare module.exports: $Exports<'webpack/lib/optimize/OccurenceOrderPlugin'>;
 }
 declare module 'webpack/lib/optimize/OccurrenceOrderPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/optimize/OccurrenceOrderPlugin'>;
@@ -1419,11 +1701,26 @@ declare module 'webpack/lib/optimize/UglifyJsPlugin.js' {
 declare module 'webpack/lib/OptionsApply.js' {
   declare module.exports: $Exports<'webpack/lib/OptionsApply'>;
 }
-declare module 'webpack/lib/OriginalSource.js' {
-  declare module.exports: $Exports<'webpack/lib/OriginalSource'>;
+declare module 'webpack/lib/OptionsDefaulter.js' {
+  declare module.exports: $Exports<'webpack/lib/OptionsDefaulter'>;
 }
 declare module 'webpack/lib/Parser.js' {
   declare module.exports: $Exports<'webpack/lib/Parser'>;
+}
+declare module 'webpack/lib/ParserHelpers.js' {
+  declare module.exports: $Exports<'webpack/lib/ParserHelpers'>;
+}
+declare module 'webpack/lib/performance/AssetsOverSizeLimitWarning.js' {
+  declare module.exports: $Exports<'webpack/lib/performance/AssetsOverSizeLimitWarning'>;
+}
+declare module 'webpack/lib/performance/EntrypointsOverSizeLimitWarning.js' {
+  declare module.exports: $Exports<'webpack/lib/performance/EntrypointsOverSizeLimitWarning'>;
+}
+declare module 'webpack/lib/performance/NoAsyncChunksWarning.js' {
+  declare module.exports: $Exports<'webpack/lib/performance/NoAsyncChunksWarning'>;
+}
+declare module 'webpack/lib/performance/SizeLimitsPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/performance/SizeLimitsPlugin'>;
 }
 declare module 'webpack/lib/PrefetchPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/PrefetchPlugin'>;
@@ -1437,9 +1734,6 @@ declare module 'webpack/lib/ProvidePlugin.js' {
 declare module 'webpack/lib/RawModule.js' {
   declare module.exports: $Exports<'webpack/lib/RawModule'>;
 }
-declare module 'webpack/lib/RawSource.js' {
-  declare module.exports: $Exports<'webpack/lib/RawSource'>;
-}
 declare module 'webpack/lib/RecordIdsPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/RecordIdsPlugin'>;
 }
@@ -1452,8 +1746,8 @@ declare module 'webpack/lib/RequestShortener.js' {
 declare module 'webpack/lib/RequireJsStuffPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/RequireJsStuffPlugin'>;
 }
-declare module 'webpack/lib/ResolverPlugin.js' {
-  declare module.exports: $Exports<'webpack/lib/ResolverPlugin'>;
+declare module 'webpack/lib/RuleSet.js' {
+  declare module.exports: $Exports<'webpack/lib/RuleSet'>;
 }
 declare module 'webpack/lib/SetVarMainTemplatePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/SetVarMainTemplatePlugin'>;
@@ -1461,17 +1755,14 @@ declare module 'webpack/lib/SetVarMainTemplatePlugin.js' {
 declare module 'webpack/lib/SingleEntryPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/SingleEntryPlugin'>;
 }
-declare module 'webpack/lib/Source.js' {
-  declare module.exports: $Exports<'webpack/lib/Source'>;
+declare module 'webpack/lib/SizeFormatHelpers.js' {
+  declare module.exports: $Exports<'webpack/lib/SizeFormatHelpers'>;
 }
 declare module 'webpack/lib/SourceMapDevToolModuleOptionsPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/SourceMapDevToolModuleOptionsPlugin'>;
 }
 declare module 'webpack/lib/SourceMapDevToolPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/SourceMapDevToolPlugin'>;
-}
-declare module 'webpack/lib/SourceMapSource.js' {
-  declare module.exports: $Exports<'webpack/lib/SourceMapSource'>;
 }
 declare module 'webpack/lib/Stats.js' {
   declare module.exports: $Exports<'webpack/lib/Stats'>;
@@ -1488,6 +1779,15 @@ declare module 'webpack/lib/UmdMainTemplatePlugin.js' {
 declare module 'webpack/lib/UnsupportedFeatureWarning.js' {
   declare module.exports: $Exports<'webpack/lib/UnsupportedFeatureWarning'>;
 }
+declare module 'webpack/lib/UseStrictPlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/UseStrictPlugin'>;
+}
+declare module 'webpack/lib/util/identifier.js' {
+  declare module.exports: $Exports<'webpack/lib/util/identifier'>;
+}
+declare module 'webpack/lib/validateSchema.js' {
+  declare module.exports: $Exports<'webpack/lib/validateSchema'>;
+}
 declare module 'webpack/lib/WarnCaseSensitiveModulesPlugin.js' {
   declare module.exports: $Exports<'webpack/lib/WarnCaseSensitiveModulesPlugin'>;
 }
@@ -1503,20 +1803,35 @@ declare module 'webpack/lib/webpack.js' {
 declare module 'webpack/lib/webpack.web.js' {
   declare module.exports: $Exports<'webpack/lib/webpack.web'>;
 }
+declare module 'webpack/lib/WebpackError.js' {
+  declare module.exports: $Exports<'webpack/lib/WebpackError'>;
+}
 declare module 'webpack/lib/WebpackOptionsApply.js' {
   declare module.exports: $Exports<'webpack/lib/WebpackOptionsApply'>;
 }
 declare module 'webpack/lib/WebpackOptionsDefaulter.js' {
   declare module.exports: $Exports<'webpack/lib/WebpackOptionsDefaulter'>;
 }
+declare module 'webpack/lib/WebpackOptionsValidationError.js' {
+  declare module.exports: $Exports<'webpack/lib/WebpackOptionsValidationError'>;
+}
 declare module 'webpack/lib/webworker/WebWorkerChunkTemplatePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/webworker/WebWorkerChunkTemplatePlugin'>;
+}
+declare module 'webpack/lib/webworker/WebWorkerHotUpdateChunkTemplatePlugin.js' {
+  declare module.exports: $Exports<'webpack/lib/webworker/WebWorkerHotUpdateChunkTemplatePlugin'>;
+}
+declare module 'webpack/lib/webworker/WebWorkerMainTemplate.runtime.js' {
+  declare module.exports: $Exports<'webpack/lib/webworker/WebWorkerMainTemplate.runtime'>;
 }
 declare module 'webpack/lib/webworker/WebWorkerMainTemplatePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/webworker/WebWorkerMainTemplatePlugin'>;
 }
 declare module 'webpack/lib/webworker/WebWorkerTemplatePlugin.js' {
   declare module.exports: $Exports<'webpack/lib/webworker/WebWorkerTemplatePlugin'>;
+}
+declare module 'webpack/schemas/ajv.absolutePath.js' {
+  declare module.exports: $Exports<'webpack/schemas/ajv.absolutePath'>;
 }
 declare module 'webpack/web_modules/node-libs-browser.js' {
   declare module.exports: $Exports<'webpack/web_modules/node-libs-browser'>;
