@@ -1,5 +1,5 @@
-// flow-typed signature: d18c37b5b4bf34fe0ea805d14d048769
-// flow-typed version: <<STUB>>/whatwg-fetch_v^2.0.1/flow_v0.41.0
+// flow-typed signature: ad26e0e609a7b647f252d9750ac45c7c
+// flow-typed version: <<STUB>>/whatwg-fetch_v^2.0.3/flow_v0.46.0
 
 /**
  * This is an autogenerated libdef stub for:
