@@ -3,10 +3,12 @@ import React from 'react';
 
 import './Spinner.scss';
 
-export default () => (
+const Spinner = () => (
   <div className="spinner">
     {_.range(1, 13).map(i =>
-      <div key={i} className={`spinner-circle${i} spinner-circle`} />
+      <div key={i} className={`spinner-circle${i} spinner-circle`} />,
     )}
   </div>
 );
+
+export default Spinner;
