@@ -1,20 +1,20 @@
 import { combineSagas } from 'util/Saga';
-import { default as userSaga } from 'routes/Auth/sagas/UserSaga';
-import { default as signUpSaga } from 'routes/Auth/sagas/SignUpSaga';
-import { default as signInSaga } from 'routes/Auth/sagas/SignInSaga';
-import { default as signOutSaga } from 'routes/Auth/sagas/SignOutSaga';
-import { default as activateAccountSaga } from 'routes/Auth/sagas/ActivateAccountSaga';
-import { default as recoverPasswordSaga } from 'routes/Auth/sagas/RecoverPasswordSaga';
-import { default as resetPasswordSaga } from 'routes/Auth/sagas/ResetPasswordSaga';
+import userSagaBinding from 'routes/Auth/sagas/UserSaga';
+import signUpSagaBinding from 'routes/Auth/sagas/SignUpSaga';
+import signInSagaBinding from 'routes/Auth/sagas/SignInSaga';
+import signOutSagaBinding from 'routes/Auth/sagas/SignOutSaga';
+import activateAccountSagaBinding from 'routes/Auth/sagas/ActivateAccountSaga';
+import recoverPasswordSagaBinding from 'routes/Auth/sagas/RecoverPasswordSaga';
+import resetPasswordSagaBinding from 'routes/Auth/sagas/ResetPasswordSaga';
 
 export default function* authSaga() {
   yield combineSagas([
-    userSaga,
-    signUpSaga,
-    signInSaga,
-    signOutSaga,
-    activateAccountSaga,
-    recoverPasswordSaga,
-    resetPasswordSaga,
+    userSagaBinding,
+    signUpSagaBinding,
+    signInSagaBinding,
+    signOutSagaBinding,
+    activateAccountSagaBinding,
+    recoverPasswordSagaBinding,
+    resetPasswordSagaBinding,
   ]);
 }
